Implemented test for ProductCreateView & ProductUpdateView & ProductDeleteView views classes
